mcp-servers/meesho/requirements.txt
mcp[server]>=0.1.0
starlette
uvicorn
httpx
python-dotenv
